pyproject.toml
[tool.poetry]
name = "concessionaria"
version = "0.1.0"
description = "Programa de concessionaria de carros que auxilia na administração dos veiculos"
authors = ["Nicholas Estrada"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
djangorestframework = "^3.13.1"
black = "^22.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
